# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Stellar Intel API",
    "version": "1.3.0",
    "description": "Intent router and anchor rate aggregation API for the Stellar Intel platform.\n\n## Public v1 surface\n\nThe stable, supported surface is namespaced under `/api/v1/...`; unversioned\nroutes are internal and may change without notice. Every v1 response follows\nthe hardening contract (see `lib/api/v1.ts`):\n\n- **Error envelope** — errors return `{ \"error\": { \"code\", \"message\", \"requestId\" } }`.\n- **Rate-limit headers** — `X-RateLimit-Limit`, `X-RateLimit-Remaining`,\n  `X-RateLimit-Reset` on every response; `Retry-After` on a `429`.\n- **Idempotency** — send an `Idempotency-Key` header on mutating endpoints\n  (e.g. `POST /api/v1/intent/offramp`); a retried request replays the original\n  response (`Idempotency-Replayed: true`) instead of creating a second intent.",
    "contact": {
      "name": "Stellar Intel",
      "url": "https://github.com/ezedike-evan/stellar-intel"
    }
  },
  "servers": [
    {
      "url": "https://stellar-intel.vercel.app/api/v1",
      "description": "Production (public v1)"
    },
    {
      "url": "https://stellar-intel.vercel.app",
      "description": "Production (internal, unversioned)"
    },
    {
      "url": "http://localhost:3000",
      "description": "Development"
    }
  ],
  "components": {
    "schemas": {
...
      "OfframpRoute": {
        "type": "object",
        "properties": {
          "anchorId": {
            "type": "string"
          },
          "anchorDomain": {
            "type": "string"
          },
          "corridorId": {
            "type": "string"
          },
          "estimatedFee": {
            "type": "string"
          },
          "estimatedReceived": {
            "type": "string"
          }
        },
        "required": [
          "anchorId",
          "anchorDomain",
          "corridorId",
          "estimatedFee",
          "estimatedReceived"
        ]
      },
      "OfframpIntentResponse": {
        "type": "object",
        "properties": {
          "route": {
            "$ref": "#/components/schemas/OfframpRoute"
          },
          "unsignedTx": {
            "type": "string",
            "description": "XDR-encoded unsigned Stellar transaction"
          },
          "quoteId": {
            "type": "string",
            "description": "Hex-encoded SHA-256 quote identifier"
          }
        },
        "required": [
          "route",
          "unsignedTx",
          "quoteId"
        ]
      },
      "ApiError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Machine-readable error code"
          },
          "message": {
            "type": "string",
            "description": "Human-readable error description"
          },
          "retryAfter": {
            "type": "number",
            "description": "Seconds until the client may retry. Only present for code === \"RATE_LIMITED\""
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "IntentRequest": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "offramp"
            ]
          },
          "sourceAsset": {
            "type": "string",
            "minLength": 1
          },
          "destinationAsset": {
            "type": "string",
            "minLength": 1
          },
          "amount": {
            "type": "string",
            "pattern": "^\\d+(\\.\\d+)?$"
          },
          "sender": {
            "type": "string",
            "minLength": 1,
            "description": "Stellar public key of the sender"
          },
          "recipient": {
            "type": "string",
            "minLength": 1,
            "description": "Destination address for the payout"
          },
          "nonce": {
            "type": "string",
            "pattern": "^[0-9a-f]{32}$/i",
            "description": "128-bit random hex, required for replay protection when the intent is signed"
          },
          "deadline": {
            "type": "string",
            "description": "RFC 3339 datetime after which the intent must not execute; required when the intent is signed"
          }
        },
        "required": [
          "type",
          "sourceAsset",
          "destinationAsset",
          "amount",
          "sender",
          "recipient"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/api/intent/offramp": {
      "post": {
        "summary": "Submit an off-ramp intent",
        "description": "Resolves an anchor route for the given asset corridor, builds an unsigned Stellar payment transaction, and returns a quote ID. Every response carries an `API-Version` header and `X-RateLimit-Limit` / `X-RateLimit-Remaining` / `X-RateLimit-Reset` headers. Send an `Idempotency-Key` header to safely retry: a repeated key within 24h replays the original response (flagged with `Idempotency-Replayed: true`) instead of re-executing the request. Only 200 and 400 responses are cached under a key; a 500 is never cached, so a retry with the same key will try again.",
        "tags": [
          "Intent"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Client-generated key. A repeated value within 24h replays the original response."
            },
            "required": false,
            "description": "Client-generated key. A repeated value within 24h replays the original response.",
            "name": "Idempotency-Key",
            "in": "header"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IntentRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Route resolved and unsigned transaction built",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfframpIntentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error or no route found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Transaction build failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/sep10/client-domain": {
      "post": {
        "summary": "Co-sign a SEP-10 client_domain challenge",
        "description": "Adds the server-held `client_domain` signature to a verified SEP-10 challenge so an anchor can attribute the session to us. The challenge is re-validated in full against the anchor’s published SIGNING_KEY (mainnet, sequence 0, anchor-signed, manage_data only, timebounds ≤ 24h) and must carry exactly one `client_domain` operation naming our domain and sourced by our signing key. This route is not a general-purpose signing oracle.",
        "tags": [
          "SEP-10"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "transaction": {
                    "type": "string",
                    "description": "Base64 SEP-10 challenge XDR from the anchor"
                  },
                  "homeDomain": {
                    "type": "string",
                    "description": "Home or service domain of a registered anchor the challenge is from"
                  }
                },
                "required": [
                  "transaction",
                  "homeDomain"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Challenge co-signed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "transaction": {
                      "type": "string",
                      "description": "The challenge XDR with our client_domain signature"
                    }
                  },
                  "required": [
                    "transaction"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request, unregistered anchor, or challenge failed validation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Server client_domain signing is not configured",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/intent/offramp": {
      "post": {
        "summary": "Submit an off-ramp intent (v1)",
        "description": "The public v1 intent endpoint. Honours `Idempotency-Key`: a retried request replays the original response with `Idempotency-Replayed: true` rather than creating a second intent.",
        "tags": [
          "Intent"
        ],
        "responses": {
          "200": {
            "description": "Intent accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OfframpIntentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}